Rename environment.ts to assignment.ts
