+++ ts-apps/decision-logic-visualizer/src/lib/eval/assignment.ts
