+++ ts-apps/decision-logic-visualizer/src/lib/eval/eval.ts
@@ -7,7 +7,7 @@ import {
   isFalseVal,
   isBoolVal,
 } from './type.js'
-import { Assignment } from './environment.js'
+import { Assignment } from './assignment.js'
 import { match } from 'ts-pattern'
 
 /**********************************************************
